+++ firedrake/_version.py
@@ -6,6 +6,7 @@
 
 # This file is released into the public domain. Generated by
 # versioneer-0.10 (https://github.com/warner/python-versioneer)
+from __future__ import absolute_import
 
 # these strings will be replaced by git during git-archive
 git_refnames = "$Format:%d$"
