+++ firedrake/expression.py
@@ -1,9 +1,10 @@
+from __future__ import absolute_import
 import numpy as np
 import ufl
 
 from pyop2 import op2
 
-import utils
+import firedrake.utils as utils
 
 
 __all__ = ['Expression']
