+++ firedrake/fiat_utils.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 from weakref import WeakKeyDictionary
 import FIAT
 import ffc
