+++ firedrake/petsc.py
@@ -1,4 +1,5 @@
 # Utility module that imports and initialises petsc4py
+from __future__ import absolute_import
 import petsc4py
 import sys
 petsc4py.init(sys.argv)
