+++ firedrake/utils.py
@@ -1,4 +1,5 @@
 # Some generic python utilities not really specific to our work.
+from __future__ import absolute_import
 from decorator import decorator
 from pyop2.utils import cached_property  # noqa: imported from here elsewhere
